+++ src/Jobs/EventJob.php
@@ -6,6 +6,7 @@ use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
+use MOIREI\EventTracking\Facades\Events;
 
 abstract class EventJob implements ShouldQueue
 {
@@ -27,5 +28,21 @@ abstract class EventJob implements ShouldQueue
         }
         $this->onQueue(config('event-tracking.queue.queue'));
         $this->through(config('event-tracking.queue.middleware', []));
+    }
+
+    /**
+     * Handle event completed.
+     *
+     * @return void
+     */
+    public function destroy()
+    {
+        // destroy all channel instances created by this job
+        foreach ($this->channels as $channel) {
+            $instance = Events::getChannel($channel);
+            if ($instance) {
+                unset($instance);
+            }
+        }
     }
 }